requirements.txt
Flask
PyPDF2
PyMuPDF
pdf2image
Pillow
python-docx
docx2pdf
